@@ -11,8 +11,10 @@
     "@testing-library/user-event": "^13.5.0",
     "dotenv": "^16.0.1",
     "emailjs-com": "^3.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "framer-motion": "^11.3.30",
+    "lucide-react": "^0.436.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.4.0",
     "react-router-dom": "^6.3.0",
     "react-router-hash-link": "^2.4.3",
